Use functools.partial to more cleanly bind loop values
See https://docs.python-guide.org/writing/gotchas/#late-binding-closures
for more information. Since the to and from functions are created
within a loop, arguments passed to them are actually all a reference
to the same variable. This means that even after they have been created,
the methods made within this loop will take on the final value of the
iterator's variable, rather than its value at the time of creation. The
functools.partial function lets us bind a specific value of i to the
a new function call which is just a wrapper around the already-generated
one.

This is an improvement over the old structure since clients can no
longer pass `i` as a keyword argument to break the behavior of the
mapping methods.